boundos:insideBaseDirEval: return true if baseDir is "/"
insideBaseDirEval would fail if baseDir was "/".
Since "/" contains anything, and can never be a symlink,
just return true, nil if baseDir is "/".

Add a simple test for this case.

Also, while we are at it, have the returned error be a lot more
informative: show the name and the base diectory; further, have it
wrap os.ErrNotExist should anyone wish to use errors.Is at some
point.

Signed-off-by: Ronald G Minnich <rminnich@gmail.com>